fixed line yml linebreaks for tickets
