+++ frontend/app/package.json
@@ -72,7 +72,6 @@
 		"@types/dateformat": "^3.0.1",
 		"@types/jest": "^24.9.1",
 		"@types/node": "^12.12.27",
-		"@types/react": "^16.9.19",
 		"@types/react-dom": "^16.9.5",
 		"bootstrap": "^4.5.0",
 		"typescript": "^3.7.5"
