+++ frontend/app/src/people/utils/renderMarkdown.tsx
@@ -1,0 +1,24 @@
+import React, { useState } from 'react';
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+
+export function renderMarkdown(markdown) {
+
+        return <ReactMarkdown children={markdown} remarkPlugins={[remarkGfm]}
+                components={{
+                        code({ node, inline, className, children, ...props }) {
+                                return (
+                                        <code className={className} {...props}>
+                                                {children}
+                                        </code>
+                                )
+                        },
+                        img({ className, ...props }) {
+                                return (
+                                        <img className={className}
+                                                style={{ width: '100%' }}
+                                                {...props} />
+                                )
+                        }
+                }} />
+}
